Accept non-base labels
Let '__dfsan_print_data_flow' traverse the entire label tree of a non-base label
rather than failing in this case. Such trees arise from the instrumentation of
loads and stores (see options 'dfsan-combine-pointer-labels-on-load' and
'dfsan-combine-pointer-labels-on-store', which are true by default), and
flattening them would involve treating loads and stores as first-class
operations. Avoid this by now for simplicity.