End tagging at the end of each latch block
End tagging at the end of each latch block rather than at the end of each latch
predecessor. This deals with the case where body basic blocks are also latches,
such as this one:

while(walker) {
  struct sphere *tmp = walker;
  walker = walker->next;
  free(tmp);
}

This will tag iterator instructions (such as "walker = walker->next" in the
example) which introduce loop-carried dependencies. We rely on iteration
recognition and the trace simplification phase to eliminate them.